pyproject.toml
[tool.poetry]
name = "chomsky-neural"
version = "0.1.0"
description = ""
authors = ["agiats <taiga98-0809@g.ecc.u-tokyo.ac.jp>"]
readme = "README.md"
packages = [{include = "chomsky_neural"}]

[tool.poetry.dependencies]
python = ">=3.8.1,<4.0"
transformers = "^4.25.1"
scikit-learn = "^1.2.0"
tqdm = "^4.64.1"
fire = "^0.5.0"
jsonlines = "^3.1.0"
pytorch-lightning = "^1.8.4.post0"
torch = "^1.13.0"
ai2-tango = "^1.1.0"
datasets = "^2.7.1"
mlflow = "^2.0.1"
tensorboard = "^2.11.0"
lightning-lite = "^1.8.4.post0"

[tool.pysen]
version = "0.10"

[tool.pysen.lint]
enable_black = true
enable_flake8 = true
enable_isort = true
enable_mypy = true
mypy_preset = "strict"
line_length = 88
py_version = "py310"

[tool.mypy]
python_version="3.10"
files = "smartgoal"
disallow_untyped_defs = true
ignore_missing_imports = true
no_implicit_optional = true
show_error_codes = true
pretty = true
allow_untyped_globals = true

[[tool.pysen.lint.mypy_targets]]
paths = ["chomsky_neural", "tests"]

[tool.poetry.group.dev.dependencies]
isort = "^5.11.1"
flake8 = "^6.0.0"
black = "^22.12.0"
mypy = "^0.991"
pysen = "^0.10.2"
pytest = "^7.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
